refactor: refactor api for management student
