+++ drizzle.config.ts
@@ -4,7 +4,7 @@
 import { defineConfig } from 'drizzle-kit';
 
 export default defineConfig({
-  schema: './db/schema.ts',
+  schema: './backend/db/schema.ts',
   out: './drizzle',
   dialect: 'sqlite'
 });
