--- drizzle/0001_unique_gressill.sql
@@ -1,17 +1,0 @@
-PRAGMA foreign_keys=OFF;--> statement-breakpoint
-CREATE TABLE `__new_users` (
-	`id` text PRIMARY KEY NOT NULL,
-	`email` text NOT NULL,
-	`name` text NOT NULL,
-	`role` text DEFAULT 'staff' NOT NULL,
-	`classroom_id` text,
-	`color` text DEFAULT '#3b82f6',
-	`deleted_at` integer,
-	FOREIGN KEY (`classroom_id`) REFERENCES `classrooms`(`id`) ON UPDATE no action ON DELETE no action
-);
---> statement-breakpoint
-INSERT INTO `__new_users`("id", "email", "name", "role", "classroom_id", "color", "deleted_at") SELECT "id", "email", "name", "role", "classroom_id", "color", "deleted_at" FROM `users`;--> statement-breakpoint
-DROP TABLE `users`;--> statement-breakpoint
-ALTER TABLE `__new_users` RENAME TO `users`;--> statement-breakpoint
-PRAGMA foreign_keys=ON;--> statement-breakpoint
-CREATE UNIQUE INDEX `users_email_unique` ON `users` (`email`);
