--- drizzle/0002_remarkable_lyja.sql
@@ -1,2 +1,0 @@
-ALTER TABLE `users` ADD `first_name` text DEFAULT '' NOT NULL;--> statement-breakpoint
-ALTER TABLE `users` ADD `last_name` text DEFAULT '' NOT NULL;
