--- drizzle/0003_lame_vertigo.sql
@@ -1,1 +1,0 @@
-ALTER TABLE `users` DROP COLUMN `name`;
