--- drizzle/0005_acoustic_night_thrasher.sql
@@ -1,2 +1,0 @@
-DROP INDEX `users_email_unique`;--> statement-breakpoint
-CREATE UNIQUE INDEX `users_email_active_unique` ON `users` (`email`) WHERE "users"."deleted_at" is null;
