--- drizzle/0006_adorable_cloak.sql
@@ -1,1 +1,0 @@
-ALTER TABLE `students` ADD `email` text NOT NULL;
